requirements.txt
requests
python-dotenv
ISStreamer